+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Aggregate projections can now be async
+
 ### Fixed
 - Bug where idempotency in read model store fails for multiple inserts/updates per event
 - Bug where projections handle events in a wrong order during a replay